refactor(team-builder): extract shared dbPokemonToFlat helper
Moves DB→PokemonSetFlat conversion from export-menu, pokemon-import-export,
and validation-hooks into a single pokemon-utils.ts file.
